+++ scripts/tasks_retrival/HPA_tasks_creation.py
@@ -1,5 +1,4 @@
 import click
-import numpy as np
 import pandas as pd
 
 from gene_benchmark.tasks import dump_task_definitions
@@ -7,6 +6,7 @@ from scripts.tasks_retrival.task_retrieval import (
     check_data_type,
     create_single_label_task,
     load_yaml_file,
+    tag_list_to_multi_label,
 )
 
 CELL_LINE = "Cell line expression cluster"
@@ -47,36 +47,6 @@ def create_tasks(data, main_task_directory):
             entities, outcomes = create_single_label_task(current_col_data)
         task_name = col.replace("/", "|")
         dump_task_definitions(entities, outcomes, main_task_directory, task_name)
-
-
-def tag_list_to_multi_label(
-    current_col_data: pd.Series, entities_name: str = "symbol", delimiter: str = ","
-) -> tuple[pd.Series, pd.DataFrame]:
-    """
-    Takes a table with entities in rows and a tag cloud of attributes in values
-      and converts into a multi label task.
-
-    Args:
-    ----
-        current_col_data (pd.Series): Series with entities as indexes and the attribute list as the values
-        entities_name (str, optional): Type of the entities. Defaults to 'symbol'.
-        delimiter (str, optional): The delimiter in the attributes cloud . Defaults to ','.
-
-    Returns:
-    -------
-        tuple[pd.Series,pd.DataFrame]: A tuple with the entities and a dta frame where each column
-        is a attribute and the values represent the assignment to each attribute
-
-    """
-    split_values_df = current_col_data.apply(
-        lambda x: [item.strip() for item in x.split(delimiter)]
-    )
-    vocab = list(set(np.concatenate(split_values_df.values)))
-    outcome_df = pd.DataFrame(0, index=split_values_df.index, columns=vocab)
-    for index in range(split_values_df.shape[0]):
-        outcome_df.iloc[index][split_values_df.iloc[index]] = 1
-    entities = pd.Series(outcome_df.index, name=entities_name)
-    return entities, outcome_df
 
 
 @click.command()
