+++ scripts/tasks_retrival/task_retrieval.py
@@ -4,6 +4,7 @@ from pathlib import Path
 from urllib.parse import urlparse
 
 import mygene
+import numpy as np
 import pandas as pd
 import requests
 from yaml import safe_load
@@ -288,3 +289,33 @@ def create_single_label_task(
     entities = pd.Series(current_col_data.index, name=entities_name)
     outcomes = pd.Series(current_col_data.values, name=outcomes_name)
     return entities, outcomes
+
+
+def tag_list_to_multi_label(
+    current_col_data: pd.Series, entities_name: str = "symbol", delimiter: str = ","
+) -> tuple[pd.Series, pd.DataFrame]:
+    """
+    Takes a table with entities in rows and a tag cloud of attributes in values
+      and converts into a multi label task.
+
+    Args:
+    ----
+        current_col_data (pd.Series): Series with entities as indexes and the attribute list as the values
+        entities_name (str, optional): Type of the entities. Defaults to 'symbol'.
+        delimiter (str, optional): The delimiter in the attributes cloud . Defaults to ','.
+
+    Returns:
+    -------
+        tuple[pd.Series,pd.DataFrame]: A tuple with the entities and a dta frame where each column
+        is a attribute and the values represent the assignment to each attribute
+
+    """
+    split_values_df = current_col_data.apply(
+        lambda x: [item.strip() for item in x.split(delimiter)]
+    )
+    vocab = list(set(np.concatenate(split_values_df.values)))
+    outcome_df = pd.DataFrame(0, index=split_values_df.index, columns=vocab)
+    for index in range(split_values_df.shape[0]):
+        outcome_df.iloc[index][split_values_df.iloc[index]] = 1
+    entities = pd.Series(outcome_df.index, name=entities_name)
+    return entities, outcome_df
